zebra: do not install the nhg for kernel/connected routes
I see this issue during below events sequencing
1. zebra creates an nhe and sets 'initial delay' flag for the nexthop
   received along with kernel/connected route and this routes is a v6
   route.
2. Later zebra receives intf_address event for the interface that
   belongs to the same nhe created above. but this is v4 event. Then
   zebra iterates through the nhe set linked to this interface and
   eventually it will end up installing this nhe in kernel

So, we install the NHG in kernel for connected/kernel routes and that
looks to be deviating from the expected behaviour.
All this happens when we receive interface event, we attempt a reinstall
for all the NHGs associated with that intf. But if the 'initial delay'
is already set for an NHG, we can skip that.
Fixing the same.

Ticket :#4541785

Signed-off-by: Krishnasamy <krishnasamyr@nvidia.com>