bgpd: Allow BGP NHT resolved nodes to go early
When BGP is using itself as a underlay, let's attempt
to notice this and in this situation process those
useful routes as early as possible.

Imagine that BGP is resolving routes through itself.
In this situation BGP is going to also have NHT for
those routes.  When this happens, have bgp notice
that there are dependent routes that resolve through
itself, and when we receive those routes move them
to the front of the processing queue to allow the
underlay to work it's way through the system as early
as is possible.

Signed-off-by: Donald Sharp <sharpd@nvidia.com>