tools: remove redundant no bgp neighbor cmds
Individual execution of all no configs related to the neighbor when the
top level neighbor command is executed, led to frr-reload.py timeout
errors in nvue.

The frr-reload.py was modified to remove all the other no command related to the neighbor config and only have "no neighbor remote-as" per neighbor when it exists.

Testing:
2025-07-18 02:21:06,147  INFO: "service integrated-vtysh-config" cannot be removed
2025-07-18 02:21:06,257  INFO: Executed "no hostname qa-sys-eth-vt30-1-2410"
2025-07-18 02:21:06,475  INFO: Executed "router bgp 65101  no neighbor swp51.301 interface remote-as external exit"
2025-07-18 02:21:09,243  INFO: Executed "router bgp 65101  no neighbor swp51.302 interface remote-as external exit"
2025-07-18 02:21:12,326  INFO: Executed "router bgp 65101  no neighbor swp51.303 interface remote-as external exit"
2025-07-18 02:21:15,635  INFO: Executed "router bgp 65101  no neighbor swp51.304 interface remote-as external exit"

Ticket: #4531892
Signed-off-by: Vijayalaxmi Basavaraj <vbasavaraj@nvidia.com>