[CUMULUS] zebra: [GR] crash while installing last route_node
Problem:

Zebra crashed while trying to install the last route node because
the dest->selected_fib of rn was pointing to corrupt memory. The crash is
because zebra saves pointer to rn while recording last route. But this
rn was invalid by the time zebra was ready to install the last route.

Fix:
Instead of saving a pointer to rn, record all the details from ctx,
required to lookup the rn from route table. Later when zebra is ready to
reinstall the last route, it can use this info to lookup the rn.

Ticket: #4546225
Signed-off-by: Pooja Jagadeesh Doijode pdoijode@nvidia.com