zebra: Abstract architectural values that may change based upon NOS
There is a set of values stored in zebra_router that are really
architectural values that may or may not change based upon the
underlying NOS being run.  Let's move them out of zebra_router
and into their own structure such that when we want to expose
these values they do not expose all of the zebra_router internals.

Signed-off-by: Donald Sharp <sharpd@nvidia.com>