lib: add a more efficient zapi init apis
memset'ing the fairly large zapi_route struct can be expensive.
reorganize the zapi_route struct so that the large fields/arrays
are at the end of the struct, and then add an init function that
initializes only a subset of the struct. Also add an init api
for the zapi_nexthop, just in case.

Signed-off-by: Mark Stapp <mjs@cisco.com>