watchfrr,nb,gRPC: [CUMULUS] [TELEMETRY] gRPC msg with empty xpath
To address the dependency of FRR to relay config when the
service starts/stops or restarts on RTS. FRR will send a  gRPC
message with empty xpath list indicating the config needs to be
replayed from RTS.
Cleaned up the frr.service file removing dependency on rts service

Testing:

BEFORE the fix the config was replayed through forcing RTS
restart on FRR re-initializing

After fix the gRPC sends the message to RTS to push the current
configuration. This has been verified with making sure RTS is not
restarted and config is replayed.

Ticket: #4615799
Signed-off-by: Ashwini Reddy <ashred@nvidia.com>