zebra [CUMULUS]: Perform fast shutdown on 1st warmboot
Earlier, system had to be rebooted twice in warm mode to achieve lossless
reboot. 1st reboot in warm mode was done to split the ASIC
tables in half. During this reboot, FRR would go down normally. 2nd reboot
in warm mode was the actual reboot where BGP would undergo GR procedure.

But this has now changed. Instead of triggering warmboot twice, new cli
`nv set system forwarding resource-mode half` is introduced to
split the ASIC table in half by restarting switchd alone. After this the
boot mode of the system will be changed to warm and any reboots after
that is expected to lossless. In order to support this, zebra needs to
do fast shutdown when the mode is set to warm in GoDown handler irrespective
of what the current mode of the system is.

This is reverting fix made in 72bbf74abc68fb7063d7af99dc3bf3f599d04613

Ticket:#4641394
Signed-off-by: Pooja Jagadeesh Doijode <pdoijode@nvidia.com>