bgpd: [SOO] trigger SOO NHG update to zebra on SOO route deletion
Issue: In some scenario route with soo  transition's from soo nhg to zebra
nhg because installed pi bitmap of soo route is not subset of route with
pi selected bitmap.

reason for this is when path info delete for soo route is triggered,
even though we calculate the select path to reflect this, the trigger
for nhg install to zebra was missing.

This issue is more prominent after clear_batch optimization where all
the soo routes and corresponding with soo arrive at faster rate.

Ticket: #4691146

Testing:
Verified on 3 clos topology with high route scale and traffic test.

Signed-off-by: Vijayalaxmi Basavaraj <vbasavaraj@nvidia.com>
Signed-off-by: Karthikeya Venkat Muppalla <kmuppalla@nvidia.com>