bgpd: [SOO] fix DEST_SOO_DEL_PENDING flag check
changes to set DEST_SOO_DEL_PENDING flag if not already set.
The existing code fails to set the flag for routes with soo
that are withdrawn. This results in stale route with soo entries.

Once the same route with soo is received again, the same dest_soo
bgp_dest_soo_hash_entry is fetched and bgp_pi_bitmap is updated,
but the dest pointer is not updated(points to the deleted dest).

During the link up scenario, if this route with soo were to move
to soo nhg, after timer expiry, the code iterates through dest
path info in - bgp_rt_with_soo_zebra_route_install. This is an
attempt to read from already freed memory(bgp_dest_soo_entry->dest)
resulting Segmentation fault.

Ticket: #4776399
Signed-off-By: Aprathi K <aprathik@nvidia.com>